python/agents/claim-adjudication-agent/pyproject.toml
[project]
name = "claim-adjudication-agent"
version = "0.1.0"
description = "AI-driven agent designed to facilitate health insurance claim processing and adjudication."
authors = [{ name = "Author", email = "author@example.com" }]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.142.0",
    "google-genai>=1.68.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "google-adk>=1.27.0",
    "google-cloud-storage>=3.10.1",
    "google-cloud-logging>=3.14.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "google-adk[eval]>=1.27.0",
    "google-cloud-aiplatform[adk,agent-engines,evaluation]>=1.142.0",
    "nest-asyncio>=1.6.0",
    "pandas>=2.2.3,<3.0.0",
    "tabulate>=0.10.0",
    "ruff>=0.15.7",
    "mypy>=1.19.1",
    "codespell>=2.4.2",
]
deployment = [
    "absl-py>=2.1.0",
]

[tool.ruff]
# Standard ruff config
line-length = 88
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["C901"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["claim_adjudication_agent"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv"]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests", "eval"]
asyncio_default_fixture_loop_scope = "function"

[tool.agent-starter-pack]
example_question = "Adjudicate the claim ID CLAIMIDX0001"

[tool.agent-starter-pack.settings]
agent_directory = "claim_adjudication_agent"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
